pyproject.toml
[project]
name = "ultron-assistant"
version = "2.0.0"
description = "Ultron v2.0 — Personal Multi-Agent AI Assistant"
requires-python = ">=3.10"
dependencies = [
    # ── Core ──────────────────────────────────────────────────────────
    "ollama>=0.3.0",
    "httpx>=0.27.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "numpy>=1.24.0",
    "psutil>=5.9.0",

    # ── Voice ─────────────────────────────────────────────────────────
    "SpeechRecognition>=3.10.0",
    "openai-whisper>=20240930",
    "edge-tts>=6.1.0",
    "pygame>=2.5.0",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.0",

    # ── Torch (pinned to avoid conflicts between whisper/easyocr) ────
    "torch>=2.3.0",
    "torchaudio>=2.3.0",

    # ── Search & Web ──────────────────────────────────────────────────
    "duckduckgo-search>=6.0.0",
    "beautifulsoup4>=4.12.0",
    "markdown>=3.6.0",

    # ── RAG & Embeddings ──────────────────────────────────────────────
    "chromadb>=0.5.0",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-ollama>=0.1.0",
    "langchain-text-splitters>=0.2.0",
    "sentence-transformers>=2.2.0",

    # ── v2.0 Multi-Agent & RPA ────────────────────────────────────────
    "openai>=1.30.0",
    "pyautogui>=0.9.54",
    "mss>=9.0.0",
    "easyocr>=1.7.0",
    "opencv-python>=4.8.0",

    # ── Memory (new unified 3-layer system) ───────────────────────────
    "tiktoken>=0.7.0",
    "aiosqlite>=0.20.0",
    "diskcache>=5.6.0",

    # ── Email Agent ───────────────────────────────────────────────────
    "aioimaplib>=1.0.1",
    "aiosmtplib>=3.0.0",

    # ── File Organizer Agent ──────────────────────────────────────────
    "watchdog>=4.0.0",

    # ── Clipboard Agent ───────────────────────────────────────────────
    "keyboard>=0.13.5",
    "pyperclip>=1.9.0",

    # ── API Security & Logging ────────────────────────────────────────
    "slowapi>=0.1.9",
    "structlog>=24.0.0",

    # ── Workspace & Agent RAG ─────────────────────────────────────────
    "playwright>=1.44.0",
    "asyncio-throttle>=1.0.0",
    "bleach>=6.1.0",
    "cssutils>=2.10.0",

    # ── Additional Provider SDKs (optional) ───────────────────────────
    "tenacity>=8.3.0",

    # ── Skill System (BÖLÜM 2) ────────────────────────────────────────
    "sentence-transformers>=2.2.0",    # already present — confirmed
    "aiosqlite>=0.20.0",               # already present — confirmed
    "google-generativeai>=0.8.0",      # Gemini SDK
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
ultron = "ultron.cli:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["ultron*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
